backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlalchemy==2.0.31
pydantic==2.8.0
prophet==1.1.5
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0
google-generativeai==0.8.0
Pillow==10.4.0
python-multipart==0.0.9
aiofiles==24.1.0
PyJWT>=2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
psycopg[binary]==3.2.3
alembic==1.14.0
